Adapted the classes for float numbers due to our library's need of that type. Also made blockTest use Position instead of Point. Removed unnessersary imports

+++ CodenameG/src/edu/chl/codenameg/model/Vector2D.java
@@ -1,23 +1,23 @@
 package edu.chl.codenameg.model;
 
 public class Vector2D {
-	private double x;
-	private double y;
+	private float x;
+	private float y;
 	
-	public Vector2D(double x,double y){
+	public Vector2D(float x,float y){
 		this.setX(x);
 		this.setY(y);
 	}
-	public double getX() {
+	public float getX() {
 		return x;
 	}
-	public void setX(double x) {
+	public void setX(float x) {
 		this.x = x;
 	}
-	public double getY() {
+	public float getY() {
 		return y;
 	}
-	public void setY(double y) {
+	public void setY(float y) {
 		this.y = y;
 	}
 	public void add(Vector2D v2d){
